backend/services/far_mesh_worker/requirements.txt
# Core dependencies for distributed inference
transformers>=4.43.0
torch>=2.3.0
accelerate>=0.28.0
bitsandbytes==0.41.1

# API client
httpx>=0.27.0
python-dotenv>=1.0.1

# GPU monitoring
pynvml>=11.5.0

# Payment tracking (lightweight)
web3>=6.15.1
